+++ NEWS.md
@@ -3,6 +3,7 @@
 
 * Fixed `extr_iris` extracts the correct number of chemicals with no repetition (#15}.
 * Fixed `extr_comtox` works when a single `download_items` different than `DTXCID` is selected (#17).
+* Feature: introduced `extr_ice_assay_names` to retrieve ice assay names  (#16).
 
 # extractox 0.1.0